app/handlers/requirements.txt
boto3
pytesseract
Pillow
